# Improve validation performance by 750x on node.js