--- a/.babelrc
+++ b/.babelrc
@@ -2,20 +2,33 @@
     "presets": [
         ["@babel/preset-env", {
             "useBuiltIns": "usage",
-            "corejs": 3
+            "modules": false,
+            "corejs": 3,
+            "loose": false,
+            "targets": {
+                "browsers": [
+                    "> 1.5%",
+                    "Opera >= 58",
+                    "Safari >= 12",
+                    "Edge >= 75",
+                    "Firefox ESR",
+                    "not dead",
+                    "not ie <= 11",
+                    "not ie_mob <= 11"
+                ]
+            }
         }]
     ],
     "plugins": [
-        "@babel/plugin-syntax-dynamic-import",
-        "@babel/plugin-syntax-import-meta",
-        "@babel/plugin-proposal-class-properties",
-        "@babel/plugin-proposal-json-strings",
-        "@babel/plugin-proposal-function-sent",
-        "@babel/plugin-proposal-export-namespace-from",
-        "@babel/plugin-proposal-numeric-separator",
-        "@babel/plugin-proposal-throw-expressions",
+        "add-module-exports",
         ["@babel/plugin-transform-runtime", {
-            "corejs": 3
+            "corejs": 3,
+            "helpers": true,
+            "regenerator": true
+        }],
+        "@babel/plugin-transform-modules-commonjs",
+        ["@babel/plugin-proposal-class-properties", {
+            "loose": false
         }]
     ]
 }