--- a/src/protocol/control_layer/broadcast_message/BroadcastMessage.js
+++ b/src/protocol/control_layer/broadcast_message/BroadcastMessage.js
@@ -3,14 +3,10 @@ import { validateIsType } from '../../../utils/validations'
 import StreamMessage from '../../message_layer/StreamMessage'
 
 export default class BroadcastMessage extends ControlMessage {
-    constructor(version, requestId, streamMessage) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, streamMessage }) {
         super(version, ControlMessage.TYPES.BroadcastMessage, requestId)
 
         validateIsType('streamMessage', streamMessage, 'StreamMessage', StreamMessage)
         this.streamMessage = streamMessage
-    }
-
-    static create(requestId, streamMessage) {
-        return new BroadcastMessage(ControlMessage.LATEST_VERSION, requestId, streamMessage)
     }
 }