--- a/src/protocol/control_layer/broadcast_message/BroadcastMessageSerializerV1.js
+++ b/src/protocol/control_layer/broadcast_message/BroadcastMessageSerializerV1.js
@@ -21,7 +21,10 @@ export default class BroadcastMessageSerializerV1 {
             serializedStreamMsg
         ] = arr
 
-        return new BroadcastMessage(version, null, StreamMessage.deserialize(serializedStreamMsg))
+        return new BroadcastMessage({
+            version,
+            streamMessage: StreamMessage.deserialize(serializedStreamMsg),
+        })
     }
 }