--- a/src/protocol/control_layer/error_response/ErrorResponse.js
+++ b/src/protocol/control_layer/error_response/ErrorResponse.js
@@ -2,7 +2,7 @@ import ControlMessage from '../ControlMessage'
 import { validateIsString } from '../../../utils/validations'
 
 export default class ErrorResponse extends ControlMessage {
-    constructor(version, requestId, errorMessage, errorCode) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, errorMessage, errorCode }) {
         super(version, ControlMessage.TYPES.ErrorResponse, requestId)
 
         validateIsString('errorMessage', errorMessage)
@@ -13,9 +13,5 @@ export default class ErrorResponse extends ControlMessage {
             validateIsString('errorCode', errorCode)
             this.errorCode = errorCode
         }
-    }
-
-    static create(requestId, errorMessage, errorCode) {
-        return new ErrorResponse(ControlMessage.LATEST_VERSION, requestId, errorMessage, errorCode)
     }
 }