--- a/src/protocol/control_layer/error_response/ErrorResponseSerializerV1.js
+++ b/src/protocol/control_layer/error_response/ErrorResponseSerializerV1.js
@@ -20,7 +20,9 @@ export default class ErrorResponseSerializerV1 {
             errorMessage,
         ] = arr
 
-        return new ErrorResponse(version, null, errorMessage)
+        return new ErrorResponse({
+            version, errorMessage
+        })
     }
 }