--- a/src/protocol/control_layer/publish_request/PublishRequest.js
+++ b/src/protocol/control_layer/publish_request/PublishRequest.js
@@ -2,7 +2,7 @@ import ControlMessage from '../ControlMessage'
 import { validateIsNotNullOrUndefined, validateIsString } from '../../../utils/validations'
 
 export default class PublishRequest extends ControlMessage {
-    constructor(version, requestId, streamMessage, sessionToken) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, streamMessage, sessionToken }) {
         super(version, ControlMessage.TYPES.PublishRequest, requestId)
 
         validateIsNotNullOrUndefined('streamMessage', streamMessage)
@@ -10,9 +10,5 @@ export default class PublishRequest extends ControlMessage {
 
         validateIsString('sessionToken', sessionToken, true)
         this.sessionToken = sessionToken
-    }
-
-    static create(requestId, streamMessage, sessionToken) {
-        return new PublishRequest(ControlMessage.LATEST_VERSION, requestId, streamMessage, sessionToken)
     }
 }