--- a/src/protocol/control_layer/resend_request/ResendFromRequestSerializerV1.js
+++ b/src/protocol/control_layer/resend_request/ResendFromRequestSerializerV1.js
@@ -34,7 +34,9 @@ export default class ResendFromRequestSerializerV1 {
             sessionToken,
         ] = arr
 
-        return new ResendFromRequest(version, requestId, streamId, streamPartition, new MessageRef(...fromMsgRefArray), publisherId, sessionToken)
+        return new ResendFromRequest({
+            version, requestId, streamId, streamPartition, fromMsgRef: new MessageRef(...fromMsgRefArray), publisherId, sessionToken
+        })
     }
 }