--- a/src/protocol/control_layer/resend_request/ResendRangeRequest.js
+++ b/src/protocol/control_layer/resend_request/ResendRangeRequest.js
@@ -9,7 +9,9 @@ import MessageRef from '../../message_layer/MessageRef'
 import ValidationError from '../../../errors/ValidationError'
 
 export default class ResendRangeRequest extends ControlMessage {
-    constructor(version, requestId, streamId, streamPartition, fromMsgRef, toMsgRef, publisherId, msgChainId, sessionToken) {
+    constructor({
+        version = ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition, fromMsgRef, toMsgRef, publisherId, msgChainId, sessionToken
+    }) {
         super(version, ControlMessage.TYPES.ResendRangeRequest, requestId)
 
         validateIsNotEmptyString('streamId', streamId)
@@ -31,10 +33,5 @@ export default class ResendRangeRequest extends ControlMessage {
         this.publisherId = publisherId
         this.msgChainId = msgChainId
         this.sessionToken = sessionToken
-    }
-
-    static create(requestId, streamId, streamPartition, fromMsgRef, toMsgRef, publisherId, msgChainId, sessionToken) {
-        return new ResendRangeRequest(ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition,
-            fromMsgRef, toMsgRef, publisherId, msgChainId, sessionToken)
     }
 }