--- a/src/protocol/control_layer/resend_request/ResendRangeRequestSerializerV2.js
+++ b/src/protocol/control_layer/resend_request/ResendRangeRequestSerializerV2.js
@@ -35,8 +35,17 @@ export default class ResendRangeRequestSerializerV2 {
             sessionToken,
         ] = arr
 
-        return new ResendRangeRequest(version, requestId, streamId, streamPartition,
-            new MessageRef(...fromMsgRefArr), new MessageRef(...toMsgRefArr), publisherId, msgChainId, sessionToken)
+        return new ResendRangeRequest({
+            version,
+            requestId,
+            streamId,
+            streamPartition,
+            fromMsgRef: new MessageRef(...fromMsgRefArr),
+            toMsgRef: new MessageRef(...toMsgRefArr),
+            publisherId,
+            msgChainId,
+            sessionToken
+        })
     }
 }