--- a/src/protocol/control_layer/resend_response/ResendResponseNoResend.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseNoResend.js
@@ -2,7 +2,7 @@ import ControlMessage from '../ControlMessage'
 import { validateIsNotEmptyString, validateIsNotNegativeInteger } from '../../../utils/validations'
 
 export default class ResendResponseNoResend extends ControlMessage {
-    constructor(version, requestId, streamId, streamPartition) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition }) {
         super(version, ControlMessage.TYPES.ResendResponseNoResend, requestId)
 
         validateIsNotEmptyString('streamId', streamId)
@@ -12,9 +12,5 @@ export default class ResendResponseNoResend extends ControlMessage {
         this.streamPartition = streamPartition
 
         validateIsNotEmptyString('requestId', requestId) // unnecessary line once V1 is dropped
-    }
-
-    static create(requestId, streamId, streamPartition) {
-        return new ResendResponseNoResend(ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition)
     }
 }