--- a/src/protocol/control_layer/resend_response/ResendResponseNoResendSerializerV1.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseNoResendSerializerV1.js
@@ -24,7 +24,9 @@ export default class ResendResponseNoResendSerializerV1 {
             requestId,
         ] = arr
 
-        return new ResendResponseNoResend(version, requestId, streamId, streamPartition)
+        return new ResendResponseNoResend({
+            version, requestId, streamId, streamPartition
+        })
     }
 }