--- a/src/protocol/control_layer/resend_response/ResendResponseNoResendSerializerV2.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseNoResendSerializerV2.js
@@ -24,7 +24,9 @@ export default class ResendResponseNoResendSerializerV2 {
             streamPartition,
         ] = arr
 
-        return new ResendResponseNoResend(version, requestId, streamId, streamPartition)
+        return new ResendResponseNoResend({
+            version, requestId, streamId, streamPartition
+        })
     }
 }