--- a/src/protocol/control_layer/resend_response/ResendResponseResendingSerializerV1.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseResendingSerializerV1.js
@@ -24,7 +24,9 @@ export default class ResendResponseResendingSerializerV1 {
             requestId,
         ] = arr
 
-        return new ResendResponseResending(version, requestId, streamId, streamPartition)
+        return new ResendResponseResending({
+            version, requestId, streamId, streamPartition
+        })
     }
 }