--- a/src/protocol/control_layer/resend_response/ResendResponseResent.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseResent.js
@@ -2,7 +2,7 @@ import ControlMessage from '../ControlMessage'
 import { validateIsNotEmptyString, validateIsNotNegativeInteger } from '../../../utils/validations'
 
 export default class ResendResponseResent extends ControlMessage {
-    constructor(version, requestId, streamId, streamPartition) {
+    constructor({ version, requestId, streamId, streamPartition }) {
         super(version, ControlMessage.TYPES.ResendResponseResent, requestId)
 
         validateIsNotEmptyString('streamId', streamId)
@@ -12,9 +12,5 @@ export default class ResendResponseResent extends ControlMessage {
         this.streamPartition = streamPartition
 
         validateIsNotEmptyString('requestId', requestId) // unnecessary line once V1 is dropped
-    }
-
-    static create(requestId, streamId, streamPartition) {
-        return new ResendResponseResent(ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition)
     }
 }