--- a/src/protocol/control_layer/resend_response/ResendResponseResentSerializerV1.js
+++ b/src/protocol/control_layer/resend_response/ResendResponseResentSerializerV1.js
@@ -24,7 +24,9 @@ export default class ResendResponseResentSerializerV1 {
             requestId,
         ] = arr
 
-        return new ResendResponseResent(version, requestId, streamId, streamPartition)
+        return new ResendResponseResent({
+            version, requestId, streamId, streamPartition
+        })
     }
 }