--- a/src/protocol/control_layer/subscribe_request/SubscribeRequestSerializerV2.js
+++ b/src/protocol/control_layer/subscribe_request/SubscribeRequestSerializerV2.js
@@ -26,7 +26,9 @@ export default class SubscribeRequestSerializerV2 {
             sessionToken,
         ] = arr
 
-        return new SubscribeRequest(version, requestId, streamId, streamPartition, sessionToken)
+        return new SubscribeRequest({
+            version, requestId, streamId, streamPartition, sessionToken
+        })
     }
 }