--- a/src/protocol/control_layer/subscribe_response/SubscribeResponse.js
+++ b/src/protocol/control_layer/subscribe_response/SubscribeResponse.js
@@ -2,7 +2,7 @@ import ControlMessage from '../ControlMessage'
 import { validateIsNotEmptyString, validateIsNotNegativeInteger } from '../../../utils/validations'
 
 export default class SubscribeResponse extends ControlMessage {
-    constructor(version, requestId, streamId, streamPartition) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition }) {
         super(version, ControlMessage.TYPES.SubscribeResponse, requestId)
 
         validateIsNotEmptyString('streamId', streamId)
@@ -10,9 +10,5 @@ export default class SubscribeResponse extends ControlMessage {
 
         this.streamId = streamId
         this.streamPartition = streamPartition
-    }
-
-    static create(requestId, streamId, streamPartition) {
-        return new SubscribeResponse(ControlMessage.LATEST_VERSION, requestId, streamId, streamPartition)
     }
 }