--- a/src/protocol/control_layer/subscribe_response/SubscribeResponseSerializerV1.js
+++ b/src/protocol/control_layer/subscribe_response/SubscribeResponseSerializerV1.js
@@ -22,7 +22,9 @@ export default class SubscribeResponseSerializerV1 {
             streamPartition,
         ] = arr
 
-        return new SubscribeResponse(version, null, streamId, streamPartition)
+        return new SubscribeResponse({
+            version, streamId, streamPartition
+        })
     }
 }