--- a/src/protocol/control_layer/subscribe_response/SubscribeResponseSerializerV2.js
+++ b/src/protocol/control_layer/subscribe_response/SubscribeResponseSerializerV2.js
@@ -24,7 +24,9 @@ export default class SubscribeResponseSerializerV2 {
             streamPartition,
         ] = arr
 
-        return new SubscribeResponse(version, requestId, streamId, streamPartition)
+        return new SubscribeResponse({
+            version, requestId, streamId, streamPartition
+        })
     }
 }