--- a/src/protocol/control_layer/unicast_message/UnicastMessage.js
+++ b/src/protocol/control_layer/unicast_message/UnicastMessage.js
@@ -3,16 +3,12 @@ import ControlMessage from '../ControlMessage'
 import StreamMessage from '../../message_layer/StreamMessage'
 
 export default class UnicastMessage extends ControlMessage {
-    constructor(version, requestId, streamMessage) {
+    constructor({ version = ControlMessage.LATEST_VERSION, requestId, streamMessage }) {
         super(version, ControlMessage.TYPES.UnicastMessage, requestId)
 
         validateIsType('streamMessage', streamMessage, 'StreamMessage', StreamMessage)
         this.streamMessage = streamMessage
 
         validateIsNotEmptyString('requestId', requestId) // unnecessary line once V1 is dropped
-    }
-
-    static create(requestId, streamMessage) {
-        return new UnicastMessage(ControlMessage.LATEST_VERSION, requestId, streamMessage)
     }
 }