--- a/src/protocol/control_layer/unicast_message/UnicastMessageSerializerV2.js
+++ b/src/protocol/control_layer/unicast_message/UnicastMessageSerializerV2.js
@@ -23,7 +23,9 @@ export default class UnicastMessageSerializerV2 {
             serializedStreamMsg
         ] = arr
 
-        return new UnicastMessage(version, requestId, StreamMessage.deserialize(serializedStreamMsg))
+        return new UnicastMessage({
+            version, requestId, streamMessage: StreamMessage.deserialize(serializedStreamMsg)
+        })
     }
 }