--- a/src/protocol/control_layer/unsubscribe_request/UnsubscribeRequestSerializerV2.js
+++ b/src/protocol/control_layer/unsubscribe_request/UnsubscribeRequestSerializerV2.js
@@ -24,7 +24,9 @@ export default class UnsubscribeRequestSerializerV2 {
             streamPartition,
         ] = arr
 
-        return new UnsubscribeRequest(version, requestId, streamId, streamPartition)
+        return new UnsubscribeRequest({
+            version, requestId, streamId, streamPartition
+        })
     }
 }