--- a/src/protocol/control_layer/unsubscribe_response/UnsubscribeResponseSerializerV1.js
+++ b/src/protocol/control_layer/unsubscribe_response/UnsubscribeResponseSerializerV1.js
@@ -22,7 +22,9 @@ export default class UnsubscribeResponseSerializerV1 {
             streamPartition,
         ] = arr
 
-        return new UnsubscribeResponse(version, null, streamId, streamPartition)
+        return new UnsubscribeResponse({
+            version, streamId, streamPartition
+        })
     }
 }