--- a/src/protocol/control_layer/unsubscribe_response/UnsubscribeResponseSerializerV2.js
+++ b/src/protocol/control_layer/unsubscribe_response/UnsubscribeResponseSerializerV2.js
@@ -24,7 +24,9 @@ export default class UnsubscribeResponseSerializerV2 {
             streamPartition,
         ] = arr
 
-        return new UnsubscribeResponse(version, requestId, streamId, streamPartition)
+        return new UnsubscribeResponse({
+            version, requestId, streamId, streamPartition
+        })
     }
 }