Benchmarking revealed that signature validation using `ethers.js` was very slow, achieving only about 4 validations per second. `ethers.js` uses the `elliptic` pure JS implementation of the elliptic curve cryptography used by both Ethereum and Streamr.

This PR implements message signing and verification using the `secp256k1` library with native bindings, delivering 3000+ validations per second on node.js. In the browser, the library claims to fall back to `elliptic`, leading me to believe that performance in the browser is not improved by these changes. How to improve performance in the browser needs to be addressed later in a separate PR. One idea there is to use a wasm implementation.

Some benchmark results:

```
      Benchmarking using ethers.js...
      Execution time: 27267 ms
      Iterations: 100
      Iterations / second: 3.6674368284006307
      rss 190.99 MB
      heapTotal 142.33 MB
      heapUsed 122 MB
      external 1.41 MB
      arrayBuffers 0.08 MB

      Benchmarking using web3.js...
      Execution time: 23567 ms
      Iterations: 100
      Iterations / second: 4.24322145372767
      rss 191.55 MB
      heapTotal 142.83 MB
      heapUsed 115.05 MB
      external 1.41 MB
      arrayBuffers 0.08 MB

      Benchmarking recover (our implementation)...
      Execution time: 2723 ms
      Iterations: 10000
      Iterations / second: 3672.4201248622844
      rss 194.21 MB
      heapTotal 143.33 MB
      heapUsed 123.1 MB
      external 1.45 MB
      arrayBuffers 0.11 MB

      Benchmarking verify (our implementation)...
      Execution time: 2643 ms
      Iterations: 10000
      Iterations / second: 3783.579265985623
      rss 194.9 MB
      heapTotal 143.83 MB
      heapUsed 110.5 MB
      external 1.41 MB
      arrayBuffers 0.08 MB
```

Notes:

- This update imports the `secp256k1` and `sha3` libraries, increasing the size of the minified browser build from `139 KiB` to  `307 KiB`. This weight increase seems significant, but OTOH it is a consequence of moving the cryptography into this shared library instead of duplicating it in the components that use this library (broker, network, JS client). For the browser, the bundle size can potentially be optimized later when the browser performance issue is addressed.

- The unit tests run in node.js, and there are no tests to verify that the browser build works. Before publishing the changes as a non-`beta` release, this should be validated. If nothing else, it needs to be manually verified when importing this package into the JS client.