@hpihkala 2020-05-26
> @hpihkala the browser crypto does support ECDSA signing and verifying so it may be possible to use the built-in crypto APIs in the browser https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/verify

IIRC the secp256k1 curve is unfortunately not available in the Web Crypto API, but I could be wrong!

> One thing to note is that all the crypto stuff in the browser is async, while all the implementation here is synchronous, so to support both will require all upstream function calls depending on these crypto methods to also become async, which they perhaps should be anyway.

Yeah could do