@timoxley 2020-06-03
> . How to improve performance in the browser needs to be addressed later in a separate PR. One idea there is to use a wasm implementation.

Also need to consider bundle size, the client is already very weighty due to the recent crypto-related additions there.